Implement private flag logic
A flag is now attributed to a hero and only visible to him. Whether or not the global cooldown is 0 determines the flag color and if the gift is shared between all or just a hero. There will still be a local cooldown for each flag picked up (independant of the global one)